requirements.txt
# Core
python-dotenv>=1.0
pyyaml>=6.0

# Deep Learning
torch>=2.1.0
torchvision>=0.16.0

# ONNX
onnx>=1.15.0
onnxruntime>=1.16.0

# Computer Vision
opencv-python-headless>=4.8.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Data Processing
numpy>=1.24.0
pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0

# Active Learning UI
streamlit>=1.28.0

# Dev
pytest>=7.0
pytest-cov>=4.0
pytest-asyncio>=0.21.0
ruff>=0.1
pre-commit>=3.0
httpx>=0.25.0
